+++ apps/web/src/components/accounts/index.ts
@@ -1,3 +1,2 @@
 export * from "./DeleteAccountModal";
-export * from "./BalanceTable";
 export * from "./ClearingAccountParticipants";